+++ num/amount.go
@@ -208,6 +208,13 @@ func (a Amount) RescaleDown(exp uint32) Amount {
 	return a
 }
 
+// RescaleRange will rescale the amount so that it fits within the provided
+// range of exponents. This is useful for ensuring that amounts are within
+// a certain range of accuracy.
+func (a Amount) RescaleRange(min, max uint32) Amount {
+	return a.RescaleUp(min).RescaleDown(max)
+}
+
 // MatchPrecision will rescale the exponent value of the amount so that it
 // matches the scale of the provided amount, but *only* if it is higher.
 func (a Amount) MatchPrecision(a2 Amount) Amount {